+++ server/config/routes.rb
@@ -4,6 +4,7 @@ Rails.application.routes.draw do
       registrations: 'v1/auth/registrations',
     }
     put '/user/disable_lock_description', to: 'users#disable_lock_description', as: :user_disableLockDescription
+    get '/follow_requested_by_me_users', to: 'users#index_of_users_follow_requested_by_me', as: :follow_requested_by_me_users
     resources :posts, only: [:create, :destroy]
     put '/posts/:id/change_lock', to: 'posts#change_lock', as: :post_changeLock
     post '/follow_requests', to: 'follow_requests#create', as: :follow_requests
