+++ server/spec/support/request_spec_helper.rb
@@ -51,4 +51,17 @@ module RequestSpecHelper
     end
     non_existemt_userid
   end
+
+  def create_mutual_follow_user(user)
+    mutual_follow_user = FactoryBot.create(:user, userid: get_non_existemt_user_id)
+    Follower.create(followed_by: user.id, follow_to: mutual_follow_user.id)
+    Follower.create(followed_by: mutual_follow_user.id, follow_to: user.id)
+    mutual_follow_user
+  end
+
+  def create_follow_requested_user_by_argument_user(user)
+    follow_requested_user = FactoryBot.create(:user)
+    user.follow_requests.create(request_to: follow_requested_user.id)
+    follow_requested_user
+  end
 end
